python/voice/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conversator-voice"
version = "0.1.0"
description = "Voice-first interface for Conversator development assistant"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "logabell" }
]
dependencies = [
    "google-genai>=0.5.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    # VoxType for Wayland audio (Linux)
    # "voxtype>=0.1.0",  # Uncomment when available
    # Discord voice
    "discord.py[voice]>=2.3.0",
    # Telegram voice
    "python-telegram-bot>=21.0",
    # Dashboard API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
]

[project.scripts]
conversator-voice = "conversator_voice.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/conversator_voice"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
